fix autosizing issue with unwrapped children
