+++ packages/expo-modules-core/ios/Core/Views/SwiftUI/AutoSizingStack.swift
@@ -28,15 +28,19 @@ extension ExpoSwiftUI {
 
     public var body: some SwiftUI.View {
       if #available(iOS 16.0, tvOS 16.0, *) {
-        content.fixedSize(horizontal: axis.contains(.horizontal), vertical: axis.contains(.vertical))
-        .onGeometryChange(for: CGSize.self) { proxy in
-          proxy.size
-        } action: {
-          let width = axis.contains(.horizontal) ? $0.width : ShadowNodeProxy.UNDEFINED_SIZE
-          let height = axis.contains(.vertical) ? $0.height : ShadowNodeProxy.UNDEFINED_SIZE
-          let size = CGSize(width: width, height: height)
-          proxy.setViewSize?(size)
-        }
+        content
+          .if(proxy.setViewSize != nil, { view in
+            view
+              .fixedSize(horizontal: axis.contains(.horizontal), vertical: axis.contains(.vertical))
+              .onGeometryChange(for: CGSize.self) { proxy in
+                proxy.size
+              } action: {
+                let width = axis.contains(.horizontal) ? $0.width : ShadowNodeProxy.UNDEFINED_SIZE
+                let height = axis.contains(.vertical) ? $0.height : ShadowNodeProxy.UNDEFINED_SIZE
+                let size = CGSize(width: width, height: height)
+                proxy.setViewSize?(size)
+              }
+          })
       } else {
         // TODO: throw a warning
         content.onAppear(perform: {
